feat(S1.2.4.1): add ordering ride endpoint
