--- backend/RideNow/src/main/java/rs/ac/uns/ftn/asd/ridenow/dto/profile/UserResponseDTO.java
@@ -1,27 +1,0 @@
-package rs.ac.uns.ftn.asd.ridenow.dto.profile;
-
-import lombok.Getter;
-import lombok.Setter;
-import rs.ac.uns.ftn.asd.ridenow.model.enums.VehicleType;
-
-@Setter
-@Getter
-public class UserResponseDTO {
-    private String email;
-    private String firstName;
-    private String lastName;
-    private boolean active;
-    private String address;
-    private String phoneNumber;
-    private String profileImage;
-
-    private String licensePlate;
-    private String vehicleModel;
-    private VehicleType vehicleType;
-    private int numberOfSeats ;
-    private boolean babyFriendly;
-    private boolean petFriendly;
-
-    public UserResponseDTO() { super(); }
-
-}
