+++ backend/RideNow/src/main/java/rs/ac/uns/ftn/asd/ridenow/model/Passanger.java
@@ -1,0 +1,19 @@
+package rs.ac.uns.ftn.asd.ridenow.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+public class Passanger extends User {
+
+    private List<Route> favoriteRoutes;
+
+    public Passanger(String email, String password, String firstName, String lastName, String phoneNumber, String address, Long id,
+                     String profileImage, boolean active, boolean blocked, List<Route> favoriteRoutes) {
+        super(email, password, firstName, lastName, phoneNumber, address,id, profileImage, active, blocked);
+        this.favoriteRoutes = favoriteRoutes;
+    }
+}
