+++ backend/RideNow/src/main/java/rs/ac/uns/ftn/asd/ridenow/service/PassengerService.java
@@ -1,0 +1,15 @@
+package rs.ac.uns.ftn.asd.ridenow.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class PassengerService {
+
+    public void addToFavorites(Long userId, Long routeId) {
+        // mock: route added to favorites
+    }
+
+    public void removeFromFavorites(Long userId, Long routeId) {
+        // mock: route removed from favorites
+    }
+}
